requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
Jinja2==2.10.1
libsass==0.19.2
markdown2==2.3.8
MarkupSafe==1.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
